backend_task/requirements.txt
cryptography==36.0.1
Cython==0.29.28
Django==4.0.2
scikit-learn==0.22.2.post1
djangorestframework==3.13.1
django-request-logging==0.7.5
django-letsencrypt==4.1.0
docutils==0.17.1
fire==0.4.0
gunicorn==20.1.0
h5py==3.6.0
HTMLParser==0.0.2
ignite==1.1.0
importlib_metadata==4.11.2
ipython==8.1.0
ipywidgets==7.6.5
Jinja2==3.0.3
joblib==1.1.0
python-jose==3.3.0
keras==2.8.0
librosa==0.9.1
loguru==0.6.0
lxml==4.8.0
matplotlib==3.5.1
numpy==1.22.2
pandas==1.4.1
peewee==3.14.9
Pillow==9.0.1
protobuf==3.19.4
psycopg2-binary==2.9.3
pyOpenSSL==22.0.0
pytest==7.0.1
python-dotenv==0.19.2
PyYAML==6.0
scikit-learn==0.22.2.post1
scipy==1.8.0
sed_eval==0.2.1
soundfile==0.10.3.post1
Sphinx==4.4.0
torch==1.10.2
tornado==6.1
tqdm==4.63.0
unicodedata2==14.0.0
wavio==0.0.4
xgboost==1.5.2
zipp==3.7.0
